# docs/swagger.yaml
definitions:
...
  mengri-flow_internal_app_dto.CreateToolRequest:
    properties:
      config:
        additionalProperties: true
        type: object
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        type: string
      type:
        type: string
      workspaceId:
        type: string
    required:
    - config
    - name
    - type
    - workspaceId
    type: object
...
  mengri-flow_internal_app_dto.ImportToolsRequest:
    properties:
      resourceId:
        type: string
    required:
    - resourceId
    type: object
...
  mengri-flow_internal_app_dto.ListToolsResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/mengri-flow_internal_app_dto.ToolResponse'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  mengri-flow_internal_app_dto.TestToolRequest:
    properties:
      input:
        additionalProperties: true
        type: object
      toolId:
        type: string
    required:
    - input
    - toolId
    type: object
  mengri-flow_internal_app_dto.ToolResponse:
    properties:
      config:
        additionalProperties: true
        type: object
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: string
      type:
        type: string
      updatedAt:
        type: string
      workspaceId:
        type: string
    type: object
  mengri-flow_internal_app_dto.ToolVersionResponse:
    properties:
      comment:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      version:
        type: integer
    type: object
...
  mengri-flow_internal_app_dto.UpdateToolRequest:
    properties:
      config:
        additionalProperties: true
        type: object
      description:
        maxLength: 500
        type: string
      name:
        maxLength: 100
        type: string
    type: object
...
  mengri-flow_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /tools:
    get:
      consumes:
      - application/json
      parameters:
      - description: 工作空间ID
        in: query
        name: workspaceId
        required: true
        type: string
      - description: 工具类型
        in: query
        name: type
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.ListToolsResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取工具列表
      tags:
      - Tool
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建工具请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mengri-flow_internal_app_dto.CreateToolRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.ToolResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 创建工具
      tags:
      - Tool
  /tools/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 工具ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.ToolResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取工具详情
      tags:
      - Tool
    put:
      consumes:
      - application/json
      parameters:
      - description: 工具ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新工具请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mengri-flow_internal_app_dto.UpdateToolRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/mengri-flow_internal_app_dto.ToolResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 更新工具
      tags:
      - Tool
  /tools/{id}/deprecate:
    post:
      consumes:
      - application/json
      parameters:
      - description: 工具ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 下线工具
      tags:
      - Tool
  /tools/{id}/publish:
    post:
      consumes:
      - application/json
      parameters:
      - description: 工具ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 发布工具
      tags:
      - Tool
  /tools/{id}/versions:
    get:
      consumes:
      - application/json
      parameters:
      - description: 工具ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/mengri-flow_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/mengri-flow_internal_app_dto.ToolVersionResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取工具版本列表
      tags:
      - Tool
  /tools/import:
    post:
      consumes:
      - application/json
      parameters:
      - description: 导入工具请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mengri-flow_internal_app_dto.ImportToolsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 批量导入工具
      tags:
      - Tool
  /tools/test:
    post:
      consumes:
      - application/json
      parameters:
      - description: 测试工具请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mengri-flow_internal_app_dto.TestToolRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/mengri-flow_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 测试工具
      tags:
      - Tool
...
swagger: "2.0"